backend/requirements.txt
# local package
-e .

# backend requirements
click
Sphinx
coverage
flake8
python-dotenv>=0.5.1
pandas
selenium
beautifulsoup4
ipykernel
cassandra-driver
pyyaml
pymilvus==2.3.2
pydantic
fastapi
uvicorn